@@ -1,32 +1,36 @@
-import { element } from 'protractor';
-import { Injectable, Inject } from '@angular/core';
-import { Observable } from "rxjs/Rx";
-import { AngularFireDatabase, FirebaseListObservable, FirebaseObjectObservable } from 'angularfire2/database';
+import 'rxjs-compat';
+import 'rxjs/add/operator/map';
+import 'rxjs/add/operator/catch';
+import { Observable } from 'rxjs/Rx';
+
+import { Injectable } from '@angular/core';
+import { AngularFireDatabase, AngularFireList } from '@angular/fire/database';
 import { StringCleaner, StringType } from '../domain/StringCleaner';
 import { DomainBuilder, DataType } from '../domain/Builder';
-import { Media } from '../domain/Media';
-import { Movie } from '../domain/Movie';
 
 @Injectable()
 export class FirebaseService {
     public cachedData: Array<{
           branchName: string,
-          data: FirebaseListObservable<any>
+          data: Observable<any>
         }> = new Array<{
           branchName: string,
-          data: FirebaseListObservable<any>
+          data: Observable<any>
         }>();
 
     constructor(
       private db: AngularFireDatabase
     ) {}
 
-    public getBranch(branchName: string): FirebaseListObservable<any> {
+    public getBranch(branchName: string): Observable<any> {
       let cachedBranch = this.cachedData.find( item => item.branchName === branchName);
       if (!cachedBranch) {
-        let item: FirebaseListObservable<any>;
 
-        item = this.db.list('/' + branchName);
+        let itemRef: AngularFireList<any>;
+        let item: Observable<any[]>;
+
+        itemRef = this.db.list('/' + branchName);
+        item = itemRef.valueChanges();
 
         // cache for future use
         let branchToCache = {branchName: branchName, data: item};
@@ -40,19 +44,14 @@ export class FirebaseService {
     public getItemFromBranch(item: string, branchName: string, itemIsRoute: boolean, buildType: DataType): Observable<any> {
       let itemString = itemIsRoute? this.getPathFromRoute(item) : item;
 
-      return this.db.list('/' + branchName,
-      {
-        query: {
-          orderByChild: 'Path',
-          equalTo: itemString
-        }
-      }).map(response => {
+      let branchItem = this.db.list('/' + branchName, ref => ref.orderByChild('Path').equalTo(itemString)).valueChanges().map(response => {
         return this.extractDomainObject(response, buildType);
       });
 
+      return branchItem;
     }
 
-    private extractDomainObject(res: FirebaseListObservable<any>, buildType: DataType): Observable<any> {
+    private extractDomainObject(res: any, buildType: DataType): Observable<any> {
       let domainObject: any;
       res.forEach(element => {
         if(element) {
@@ -72,15 +71,4 @@ export class FirebaseService {
       return path;
     }
 
-    /**
-     * Handles the error message if the http request fails for some reason.
-     * @param error The error coming from the http request
-     * @returns {ErrorObservable} Observable exception thrown that will bubble up to the subscriber.
-     */
-    private handleError (error: any) {
-        let errMsg = (error.message) ? error.message :
-            error.status ? `${error.status} - ${error.statusText}` : 'Server error';
-        console.error(errMsg);
-        return Observable.throw(errMsg);
-    }
 }
