@@ -5,8 +5,8 @@ import { RouterTestingModule }   from '@angular/router/testing';
 import { FormsModule } from '@angular/forms';
 import { NgModule } from '@angular/core';
 import { environment } from '../../../../environments/environment';
-import { AngularFireModule } from 'angularfire2';
-import { AngularFireDatabaseModule } from 'angularfire2/database';
+import { AngularFireModule } from '@angular/fire';
+import { AngularFireDatabaseModule } from '@angular/fire/database';
 import { NavComponent } from './nav.component';
 import { SearchResultTabComponent } from '../../sub-pages/search-result-tab/search-result-tab.component';
 import { FirebaseService } from '../../../services/firebase.service';
