@@ -7,8 +7,8 @@ import { FormsModule } from '@angular/forms';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { MediaComponent } from './media.component';
 import { environment } from '../../../../environments/environment';
-import { AngularFireModule } from 'angularfire2';
-import { AngularFireDatabaseModule } from 'angularfire2/database';
+import { AngularFireModule } from '@angular/fire';
+import { AngularFireDatabaseModule } from '@angular/fire/database';
 import { FirebaseService } from '../../../services/firebase.service';
 import { ReleaseYearSortPipe } from 'app/pipes/release-year-sort.pipe';
 
