@@ -3,8 +3,8 @@ import { RouterTestingModule } from '@angular/router/testing';
 import { FormsModule } from '@angular/forms';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
 import { environment } from '../../../../environments/environment';
-import { AngularFireModule } from 'angularfire2';
-import { AngularFireDatabaseModule } from 'angularfire2/database';
+import { AngularFireModule } from '@angular/fire';
+import { AngularFireDatabaseModule } from '@angular/fire/database';
 import { MoviedetailsComponent } from './moviedetails.component';
 import { MediaGalleryComponent } from '../../sub-pages/media-gallery/media-gallery.component';
 import { FirebaseService } from '../../../services/firebase.service';
